move metric definitions to utils
